Move parameter resolver to annotation package
Move parameter resolver to annotation package

#3401
